+++ agixt/Prompts.py
@@ -1,81 +1,4 @@
 from DBConnection import Prompt, PromptCategory, Argument, session
-import os
-
-
-def import_prompts():
-    # Add default category if it doesn't exist
-    default_category = session.query(PromptCategory).filter_by(name="Default").first()
-
-    if not default_category:
-        default_category = PromptCategory(
-            name="Default", description="Default category"
-        )
-        session.add(default_category)
-        session.commit()
-        print("Adding Default prompt category")
-
-    # Get all prompt files in the specified folder
-    for root, dirs, files in os.walk("prompts"):
-        for file in files:
-            prompt_category = None
-            if root != "prompts":
-                # Use subfolder name as the prompt category
-                category_name = os.path.basename(root)
-                prompt_category = (
-                    session.query(PromptCategory).filter_by(name=category_name).first()
-                )
-                if not prompt_category:
-                    prompt_category = PromptCategory(
-                        name=category_name, description=f"{category_name} category"
-                    )
-                    session.add(prompt_category)
-                    session.commit()
-            else:
-                # Assign to "Uncategorized" category if prompt is in the root folder
-                prompt_category = default_category
-
-            # Read the prompt content from the file
-            with open(os.path.join(root, file), "r") as f:
-                prompt_content = f.read()
-
-            # Check if prompt with the same name and category already exists
-            prompt_name = os.path.splitext(file)[0]
-            prompt = (
-                session.query(Prompt)
-                .filter_by(name=prompt_name, prompt_category=prompt_category)
-                .first()
-            )
-            prompt_args = []
-            for word in prompt_content.split():
-                if word.startswith("{") and word.endswith("}"):
-                    prompt_args.append(word[1:-1])
-            if not prompt:
-                # Create the prompt entry in the database
-                prompt = Prompt(
-                    name=prompt_name,
-                    description="",
-                    content=prompt_content,
-                    prompt_category=prompt_category,
-                )
-                session.add(prompt)
-                session.commit()
-                print(f"Adding prompt: {prompt_name}")
-
-            # Populate prompt arguments
-            for arg in prompt_args:
-                if (
-                    session.query(Argument)
-                    .filter_by(prompt_id=prompt.id, name=arg)
-                    .first()
-                ):
-                    continue
-                argument = Argument(
-                    prompt_id=prompt.id,
-                    name=arg,
-                )
-                session.add(argument)
-                session.commit()
-                print(f"Adding prompt argument: {arg} for {prompt_name}")
 
 
 class Prompts:
